Version Final

+++ app/src/main/java/com/openclassrooms/entrevoisins/service/NeighbourApiService.java
@@ -32,9 +32,10 @@ public interface NeighbourApiService {
      */
     void createNeighbour(Neighbour neighbour);
 
-    void addFavorisNeighbour(int idNeighbour);
+    void addFavorisNeighbour(long idNeighbour);
 
     ArrayList<Neighbour> getFavorisNeighbour();
+
 
 
 }
